+++ src/main/java/roomescape/controller/ReservationsController.java
@@ -1,7 +1,6 @@
-package roomescape;
+package roomescape.controller;
 
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationDto;
@@ -11,20 +10,10 @@ import java.util.List;
 import java.util.Objects;
 import java.util.concurrent.atomic.AtomicLong;
 
-@Controller
-public class RoomescapeController {
+@RestController
+public class ReservationsController {
     private final List<Reservation> reservations = new ArrayList<>();
     private final AtomicLong id = new AtomicLong(1);
-
-    @GetMapping("admin")
-    public String admin() {
-        return "/admin/index";
-    }
-
-    @GetMapping("admin/reservation")
-    public String reservation() {
-        return "/admin/reservation-legacy";
-    }
 
     @GetMapping("reservations")
     public ResponseEntity<List<Reservation>> read() {
